dom/media/mediasource/test/test_SeekToLastFrame_mp4.html

<!DOCTYPE html>
<html><head>
<meta http-equiv="content-type" content="text/html; charset=windows-1252">
  <title>MSE: Can seek to last frame</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test"><script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(function(ms, el) {
  el.controls = true;
  once(ms, "sourceopen").then(function() {
    ok(true, "Receive a sourceopen event");
    const sb = ms.addSourceBuffer("video/mp4");
    fetchAndLoad(sb, "bipbop/bipbop_480_624kbps-video", [ "init" ], ".mp4")
    .then(fetchAndLoad.bind(null, sb, "bipbop/bipbop_480_624kbps-video", range(1, 3), ".m4s"))
    .then(function() {
      el.play();
      // let seek to the last audio frame.
      el.currentTime = 1.532517;
      return once(el, "seeked");
    })
    .then(function() {
      ok(true, "seek completed");
      ms.endOfStream();
      return once(el, "ended");
    }).then(function() {
      SimpleTest.finish();
    });
  });
});

</script>
</pre>
</body>
</html>
